refactor: extract type coercion to seperate file
